idk what i did its been so long all i know was there was a bug in navigation for adminviewconplaint

+++ Meeting Scheduling System/VeiwUnTrusted/EditPayment.cs
@@ -28,7 +28,7 @@ namespace Meeting_Scheduling_System.VeiwUnTrusted
 
         private void ButtonSubmit_Click(object sender, EventArgs e)
         {
-            int cardNumber = Int32.Parse(MaskedCardNumField.Text);
+            long cardNumber = Int64.Parse(MaskedCardNumField.Text);
             int ccv = Int32.Parse(MaskedCCVField.Text);
 
             AccountController.CurrentAccount.PaymentMethod.CardNumber = cardNumber;
